requirements.txt
pyside6
openai
python-dotenv

from dotenv import load_dotenv

load_dotenv()

